requirements.txt
# Core Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9

# AI/ML
numpy==1.26.2
scikit-learn==1.3.2
transformers==4.35.2
torch==2.1.1

# API Integration
requests==2.31.0
python-dotenv==1.0.0

# Authentication & Security
PyJWT==2.8.0
bcrypt==4.1.1
cryptography==41.0.7

# Date/Time
python-dateutil==2.8.2
pytz==2023.3

# Communication
twilio==8.11.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Utilities
jsonschema==4.20.0
pyyaml==6.0.1
